Füge Widget für Task-Force Tickets hinzu
Füge OpenTaskForceIssues hinzu

Hole Sprint-Tickets einmalig und füge TaskForce Tickets hinzu

Fix: Fehlende Referenz

(draft) Füge Ticket Layout hinzu

(remove) Mal die TF Tickets ins Dashboard integriert

Fix: Methodname

Füge aktuelle Jira Issue Icons hinzu

Aktualisiere Darstellungsgrößen

Formatiere Datum um

Entferne ungenutzten Output

Nutze Collection Data-Binding

Nutze immer volle Breite für ein Ticket

Korrigiere Task Force Ticket Auswahl

Füge Ansicht für "Keine Offene Task Force" hinzu

Aktualisiere Styling

Entferne ungenutzte Jira Grafiken
