+++ dashboards/produktionbauen.erb
@@ -30,6 +30,9 @@ $(function() {
   <li data-row="1" data-col="1" data-sizex="1" data-sizey="1">
     <div data-id="currentVersion" data-view="CurrentVersion"></div>
   </li>
+  <li data-row="1" data-col="1" data-sizex="1" data-sizey="2">
+    <div data-id="openTaskForceIssues" data-view="JiraOpenTaskForceIssues"></div>
+  </li>
   <li data-row="2" data-col="4" data-sizex="1" data-sizey="1">
     <div class="widget static-prosoz-logo" data-id="prosozLogo" data-view="prosozLogo">
       <img src="PROSOZ_Herten_Logo.svg" />
