+++ widgets/jira_open_task_force_issues/jira_open_task_force_issues.coffee
@@ -1,0 +1,6 @@
+class Dashing.JiraOpenTaskForceIssues extends Dashing.Widget
+
+ ready: ->
+  # nothing to do.
+ onData: (data) ->
+  # nothing to do.
